Refactor : 파이차트 리팩토링
